Add files via upload
Within this folder you will updated html screens along with their supporting CSS style guides that make up the three pages. 1. Index (main task page) 2. Login 3. Sign Up All three pages are interlinked as shown in the demo video before. In addition the html file titled opentask.html represents the open version of the accordion list which would be generated as a result of the user selecting the add button or the drop down on the task. That page is not interlinked with anything but is connected with its style sheet and guides titled opentaskguide and opentaskstyle